requirements.txt
boto3==1.38.8
botocore==1.38.8
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.1
idna==3.10
jmespath==1.0.1
playwright==1.52.0
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
